LayoutTests/http/tests/media/hls/hls-accessiblity-describes-video-menu-expected.txt

This tests the track menu when a file has a descriptive audio track.

This test only runs in DRT!


EVENT: error
PASS: We are using the apple idiom

EVENT: canplaythrough

Test when track menu is hidden

PASS: Menu name is 'Track Menu'
PASS: Menu does not exist

Test with visible menu

PASS: Menu name is 'Track Menu'
PASS: Menu exists
PASS: id is 'audioAndTextTrackMenu'
PASS: Menu element contains <ul>
PASS: DVS audio menu item name contains track name from manifest
PASS: DVS menu item name has been augmented
PASS: Standard audio menu item name is the track name from manifest

Testing finished.
